+++ traffic_ops/v4-client/async_status.go
@@ -1,0 +1,34 @@
+package client
+
+/*
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+import (
+	"fmt"
+
+	"github.com/apache/trafficcontrol/lib/go-tc"
+	"github.com/apache/trafficcontrol/traffic_ops/toclientlib"
+)
+
+// apiAsyncStatus is the API version-relative path for the /async_status/{id} API endpoint.
+const apiAsyncStatus = "/async_status/%d"
+
+// GetAsyncStatus gets an async_status from Traffic Ops.
+func (to *Session) GetAsyncStatus(id int, opts RequestOptions) (tc.AsyncStatusResponse, toclientlib.ReqInf, error) {
+	route := fmt.Sprintf(apiAsyncStatus, id)
+	var data tc.AsyncStatusResponse
+	reqInf, err := to.get(route, opts, &data)
+	return data, reqInf, err
+}